app/requirements.txt
load_dotenv
streamlit
streamlit_option_menu
streamlit-lottie
tiktoken
faiss-cpu
pinecone
pinecone-client
langchain-openai
langchain-core

